python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-hydro"
version = "1.0.0"
description = "AI-Hydro: Validated hydrological tools for agentic research platforms"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [
    {name = "Mohammad Galib", email = "mgalib@purdue.edu"}
]
keywords = ["hydrology", "watershed", "geospatial", "MCP", "agentic", "CAMELS", "FAIR"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Hydrology",
    "Topic :: Scientific/Engineering :: GIS",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

# ── Core dependencies (minimal — always installed) ──────────────────────────
# These are required for HydroResult types and basic functionality.
# Heavy geo/hydro deps are in optional extras so researchers can install
# only what they need.
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "python-dateutil>=2.8.2",
    "pyyaml>=6.0.0",
    "hydrofunctions>=0.2.0",
]

# ── Optional extras (semantic layers) ────────────────────────────────────────
[project.optional-dependencies]

# Foundation geo stack (shared by data + analysis)
geo = [
    "geopandas>=0.13.0",
    "shapely>=2.0.0",
    "pyproj>=3.6.0",
]

# Data retrieval: streamflow, forcing, land cover, soil, CAMELS
data = [
    "ai-hydro[geo]",
    "pynhd>=0.16.0",
    "pygeohydro>=0.16.0",
    "pygridmet>=0.15.0",
    "xarray>=2023.0.0",
]

# Analysis: watershed, signatures, TWI, geomorphic, curve number
analysis = [
    "ai-hydro[geo]",
    "rioxarray>=0.15.0",
    "xarray>=2023.0.0",
    "rasterio>=1.3.0",
    "py3dep>=0.15.0",
    "xrspatial>=0.3.0",
    "pysheds>=0.3.0",
    "scipy>=1.11.0",
]

# Modelling: HBV-light (PyTorch), NeuralHydrology LSTM
modelling = [
    "torch>=2.0.0",
]

# CAMELS attribute extraction (wraps camels-attrs package)
camels = [
    "camels-attrs>=0.1.0",
]

# MCP server: expose all tools as MCP endpoints
mcp = [
    "fastmcp>=0.1.0",
]

# RAG concept search (optional intelligence layer)
rag = [
    "sentence-transformers>=2.3.0",
    "chromadb>=0.4.0",
    "pyyaml>=6.0.0",
]

# Visualization helpers
viz = [
    "matplotlib>=3.7.0",
    "plotly>=5.18.0",
    "folium>=0.15.0",
]

# Full install: all tools + MCP server (no RAG)
all = [
    "ai-hydro[data,analysis,modelling,camels,mcp,viz]",
]

# Development
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "fastmcp>=0.1.0",
]

# Legacy aliases (backward compat for existing install scripts)
watershed = ["ai-hydro[data,analysis]"]
hydrology = ["ai-hydro[data]"]
geomorphic = ["ai-hydro[analysis]"]
forcing = ["ai-hydro[data]"]

# ── Entry points ─────────────────────────────────────────────────────────────
[project.scripts]
# Run the MCP server directly
aihydro-mcp = "mcp_server:main"

# Community plugin tool entry point namespace.
# Third-party packages add tools in their own pyproject.toml:
#   [project.entry-points."aihydro.tools"]
#   flood_frequency = "my_package.tools:flood_freq_tool"
# Then discover_tools() in ai_hydro.mcp.registry picks them up.
[project.entry-points."aihydro.tools"]

# ── URLs ──────────────────────────────────────────────────────────────────────
[project.urls]
Homepage = "https://github.com/galib9690/AI-Hydro"
Documentation = "https://github.com/galib9690/AI-Hydro/wiki"
Repository = "https://github.com/galib9690/AI-Hydro"
Issues = "https://github.com/galib9690/AI-Hydro/issues"

# ── Package discovery ─────────────────────────────────────────────────────────
[tool.setuptools.packages.find]
where = ["."]
include = ["ai_hydro*"]

[tool.setuptools.package-data]
ai_hydro = [
    "knowledge/**/*.json",
    "knowledge/**/*.yaml",
    "knowledge/**/*.yml",
]

# ── Pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
markers = [
    "live: end-to-end tests requiring internet and full deps (deselect with -m 'not live')",
]

# ── Ruff linting ──────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]
